impl(angular-version): Angular version
Implement the book search application using Angular. This code works
with Angular 2.0-RC5 and Angular 2.0.0.

The functionality in this version is identical to the functionality in
the React version of the book search application. The Angular version
uses Angular components, but Redux-related code, such as reducers, the
store, state history, etc. are unchanged from the React version.